requirements.txt
# knbase
PyYAML==6.0.2

# knbase-pdf-parser
pikepdf==9.2.1
pdfplumber==0.11.4
shapely==2.0.6

# knbase-index
spacy==3.8.2
langid==1.1.6
chromadb==0.5.20
sentence-transformers==3.0.1

# app
flask==3.1.0